Spring parameters now are double

+++ Source/Reverb.h
@@ -3,27 +3,27 @@
 #include <vector>
 
 struct Vector {
-    float x, y;
+    double x, y;
 
-    Vector(float x, float y);
+    Vector(double x, double y);
 
-    float magnitude() const;
+    double magnitude() const;
     Vector normalize() const;
 
     Vector operator+(const Vector& other) const;
     Vector operator-(const Vector& other) const;
-    Vector operator*(const float other) const;
-    Vector operator/(const float other) const;
+    Vector operator*(const double other) const;
+    Vector operator/(const double other) const;
 
     void operator+=(const Vector& other);
     void operator-=(const Vector& other);
-    void operator*=(const float other);
-    void operator/=(const float other);
+    void operator*=(const double other);
+    void operator/=(const double other);
 };
 
 class Particle {
 public:
-    Particle(float x, float y, float mass, bool locked);
+    Particle(double x, double y, double mass, bool locked);
 
     void applyForce(const Vector& force);
     void update();
@@ -34,32 +34,32 @@ private:
     Vector position;
     Vector velocity;
     Vector acceleration;
-    float mass;
+    double mass;
     bool locked;
 };
 
 class Spring {
 public:
-    Spring(Particle& a, Particle& b, const float k, const float restLength);
-    Spring(Particle&& a, Particle&& b, const float k, const float restLength) = delete;
+    Spring(Particle& a, Particle& b, const double k, const double restLength);
+    Spring(Particle&& a, Particle&& b, const double k, const double restLength) = delete;
 
     void update();
 
 private:
-    const float k;
-    const float restLength;
+    const double k;
+    const double restLength;
     Particle& a;
     Particle& b;
 };
 
 class SpringArray {
 public:
-    SpringArray(const size_t nodes, const float mass, const float k, const float length);
+    SpringArray(const size_t nodes, const double mass, const double k, const double length);
 
-    void applySample(const float force);
+    void applySample(const double force);
     void update();
     
-    float getSample();
+    double getSample();
 private:
     std::vector<Particle> particles;
     std::vector<Spring> springs;
